requirements/dev-requirements.txt
-r requirements.txt

autopep8==1.2.2
beautifulsoup4==4.4.1
django-debug-toolbar==1.5
django-extensions==1.6.1
django-easy-pdf
django-floppyforms==1.6.1
easy-thumbnails==2.3
flake8==2.5.4
ipdb==0.9.0
pytest-ipdb==0.1-prerelease2
xhtml2pdf
Weasyprint
reportlab
argparse==1.2.1

